Unify partition functions.
Extend the definition of partition function from mapping one tuple to
one partition to mapping one tuple to multiple destinations. In the
context of IPC, a destination corresponds to a channel determined by
a pair of (operatorId, workerId). The number of destinations of a
partition function is then (mostly) set by the producer.

Add broadcast, rewrite several partition functions. This design also
simplifies the producer partition-and-write-to-channels logic a lot.